Database has been fully corrected

+++ Models/Effeciency.cs
@@ -2,7 +2,7 @@ namespace Hackathon.Models;
 
 public class Efficiency {
     public int Id { get; set; }
-    public double Tons { get; set; }
+    public int Tons { get; set; }
 
     public override string ToString()
     {
