+++ Models/Planting.cs
@@ -4,7 +4,7 @@ public class Planting {
     public int Id { get; set; }
     public double AppliedHA { get; set; }
     public DateTime Date { get; set; }
-    public double PlantPopulation { get; set; }
+    public int PlantPopulation { get; set; }
 
     public override string ToString()
     {
